% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{compute_expected_U_m}
\alias{compute_expected_U_m}
\title{Compute the expected utility, expected squared utility, and expected medical spending with respect to the preference parameters.}
\usage{
compute_expected_U_m(
  data_set,
  param,
  n_draw,
  default_sigma = 0.5,
  taylor_order
)
}
\arguments{
\item{data_set}{This is a list that contains all pre-computed elements of each household .}

\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}

\item{n_draw}{this is number of Gauss Laguerre points.}

\item{default_sigma}{is between 0 and 1. Default to 0.5. represents the value of sigma_theta that is used for Taylor second-order approximation. sigma_theta is computed as default_sigma * exp(param$sigma_thetabar)}

\item{taylor_order}{an integer, equal to the order of Taylor approximation used to approximate exp(-x^2/sigma^2) around (sigma^2 = sigma_0^2).}
}
\value{
a list that includes the draws of U, U squared, and m. These draws are list with length equal to the order of the Taylor approximation. Each component of U and U2 is a matrix of dimension the number of bundles x length of thetabar. Each component of m is a matrix with dimension equal to length of thetabar x hhsize.
}
\description{
Compute the expected utility, expected squared utility, and expected medical spending with respect to the preference parameters.
}
\examples{
mini_data = household_draw_theta_kappa_Rdraw(1, sample_data_and_parameter$param, 1000, 10);
compute_expected_U_m(mini_data, sample_data_and_parameter$param, n_draw, default_sigma = 0.5, taylor_order=2)
}
