% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{counterfactual_bundling}
\alias{counterfactual_bundling}
\title{Function used in the constraint_function argument of counterfactual_household_draw_theta_kappa}
\usage{
counterfactual_bundling(x)
}
\arguments{
\item{x}{a vector of size n + 1, where n is the size of eligible household members}
}
\value{
a vector of size 2 (all is insured or none is insured)
}
\description{
Function used in the constraint_function argument of counterfactual_household_draw_theta_kappa
}
\examples{
counterfactual_bundling(c(1:6))

}
