% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{counterfactual_household_draw_theta_kappa_Rdraw}
\alias{counterfactual_household_draw_theta_kappa_Rdraw}
\title{Compute the counterfactual draws (preference parameters independent). This is different from household_draw_theta_kappa_Rdraw because the draws of thetabar
is only conducted once, and there is no integration over thetabar. Also, the insurance bundles are not one-person deviation from the observed bundle, but
all possible bundles.}
\usage{
counterfactual_household_draw_theta_kappa_Rdraw(
  hh_index,
  param,
  n_draw_halton = 1000,
  n_draw_gauss = 10,
  sick_parameters,
  xi_parameters,
  household_random_factor,
  individual_factor
)
}
\arguments{
\item{hh_index}{index (from object data_hh_list) of a particular household}

\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}

\item{n_draw_halton}{this is the number of halton draws, default to 1000}

\item{n_draw_gauss}{this is number of Gauss  points.}

\item{household_random_factor}{is the draw of the household's common shock that generates the within-household correlation in health risks, should come from a N(0,1) distribution}

\item{individual_factor}{is a vector (size HHsize) of the draws of the individual members' health shock (should come from a N(0,1) distribution)}
}
\value{
a list that includes the draws of household-related objects,taking into account the sick parameters and the distribution of coverage. This does not take into account estimated preference parameters or unconditional distribution of health shocks. See \code{compute_expected_U_m} for the draws post-estimation of preference parameters and health shocks distribution.
}
\description{
Compute the counterfactual draws (preference parameters independent). This is different from household_draw_theta_kappa_Rdraw because the draws of thetabar
is only conducted once, and there is no integration over thetabar. Also, the insurance bundles are not one-person deviation from the observed bundle, but
all possible bundles.
}
\examples{
individual_factor = rnorm(nrow(sample_data_and_parameter$data$data))
counterfactual_household_draw_theta_kappa_Rdraw(1, sample_data_and_parameter$param, 1000, 10, sick_parameters_sample, xi_parameters_sample, rnorm(1), individual_factor)

}
