% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{moment_eligible_hh}
\alias{moment_eligible_hh}
\title{Compute the moments of eligible households}
\usage{
moment_eligible_hh(
  data_set,
  param,
  deriv_tol = 0.001,
  default_sigma,
  taylor_order = 3
)
}
\arguments{
\item{data_set}{This is a list that contains all pre-computed elements of each household .}

\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}

\item{deriv_tol}{Default set to 1e-3. The tolerance level for the numerical computation of derivative.}

\item{default_sigma}{Is a number between 0 and 1 to use in the Taylor approximation}

\item{taylor_order}{is the order of Taylor approximation}
}
\value{
a list. The first element is expected value of oop conditional on the observed insurance status being optimal, the second element is the expected value of oop squared conditional on the observed insurance status being optimal, the third element is the probability of observed insurance being optimal. The next 3 elements are the derivative w.r.t the coefficients being optimized of the first 3 elements.
}
\description{
Compute the moments of eligible households
}
\examples{
mini_data = household_draw_theta_kappa_Rdraw(1, sample_data_and_parameter$param, 1000, 10);
mini_data = compute_expected_U_m(mini_data, sample_data_and_parameter$param, 
n_draw, default_sigma = 0.5, taylor_order=4)
moment_eligible_hh(mini_data, sample_data_and_parameter$param, deriv_tol=1e-3, default_sigma = 0.5, taylor_order=4)

}
